Fix convention error
