+++ app/Models/Admin.php
@@ -3,9 +3,13 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
 
 class Admin extends Model
 {
+    use SoftDeletes;
+
     /**
      * The attributes that are mass assignable.
      *
@@ -22,5 +26,9 @@ class Admin extends Model
      */
     protected $hidden = [
         'password', 'remember_token',
+    ];
+
+    protected $date = [
+        'deleted_at',
     ];
 }
