+++ app/Models/Salary.php
@@ -3,9 +3,13 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
 
 class Salary extends Model
 {
+    use SoftDeletes;
+
     protected $table = "salaries";
 
     protected $fillable = [
@@ -14,6 +18,6 @@ class Salary extends Model
 
     public function user()
     {
-        return $this->belongsTo('App\User');
+        return $this->belongsTo(User::class);
     }
 }
